LongMemory

Documentation for LongMemory.

LongMemory.arfigen

LongMemory.arfimagen

LongMemory.csa_llk

LongMemory.csa_var_matrix

LongMemory.csa_var_vals

LongMemory.csadiff

LongMemory.csagen

LongMemory.csagen

LongMemory.csamle_est

LongMemory.edmgen

LongMemory.exact_whittle_est

LongMemory.exact_whittle_llk

LongMemory.fi

LongMemory.fi_llk

LongMemory.fi_survival_probs

LongMemory.fi_var_matrix

LongMemory.fi_var_vals

LongMemory.fimle_est

LongMemory.fracdiff

LongMemory.fracdiff

LongMemory.gph_est

LongMemory.gph_est_variance

LongMemory.gph_est_variance

LongMemory.my_toeplitz

LongMemory.periodogram

LongMemory.whittle_est

LongMemory.whittle_est_variance

LongMemory.whittle_est_variance

LongMemory.whittle_llk

LongMemory.arfigen — Method

arfigen(T::Int, μ::Real, AR::Array, d::Real; σ=1)

Generate a time series with long memory parameter d and length T using the ARFIMA(p,d,0) model.

Arguments

T::Int: length of the time series

μ::Float64: mean of the time series

AR::Array: AR coefficients

d::Float64: fractional difference parameter

Optional arguments

σ::Float64: standard deviation of the time series

Output

x::Vector: time series

Notes

The code is inspired by the function dgp_arfima.m by Carlos Vladimir Rodríguez Caballero (2023)

Examples

julia> arfigen(100, 0, [0.2; -0.5], 0.4])

LongMemory.arfimagen — Method

arfimagen(T::Int, μ::Real, AR::Array, d::Real, MA::Array; σ=1)

Generate a time series with long memory parameter d and length T using the ARFIMA(p,d,q) model.

Arguments

T::Int: length of the time series

μ::Float64: mean of the time series

AR::Array: AR coefficients

d::Float64: fractional difference parameter

MA::Array: MA coefficients

Optional arguments

σ::Float64: standard deviation of the time series

Output

x::Vector: time series

Notes

The code is inspired by the function dgp_arfima.m by Carlos Vladimir Rodríguez Caballero (2023)

Examples

julia> arfimagen(100, 0, [0.2; -0.5], 0.4, [-0.3; 0.1]])

LongMemory.csa_llk — Method

csa_llk(p::Real, q::Real, x::Array)

Computes the log-likelihood of the CSA process with parameters p and q given the data x.

Arguments

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

x::Array: The data.

Notes

This function computes the concentrated log-likelihood function of the CSA process with parameters p and q given the data x.

Examples

julia> csa_llk(1.4, 1.8, randn(100,1))

LongMemory.csa_var_matrix — Method

csa_var_matrix(T::Int, d::Real)

Constructs the autocovariance matrix of the CSA process with parametersp and q at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

Examples

julia> fi_var_matrix(10, 1.4, 1.8)

LongMemory.csa_var_vals — Method

csa_var_vals(T::Int, p::Real, q::Real)

Computes the autocovariance function of the CSA process with parameters p and q at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

Notes

This function uses the recursive formula for the autocovariance function of the CSA process.

Examples

julia> csa_var_vals(20, 0.4, 0.6)

LongMemory.csadiff — Method

csadiff(x,p,q)

Generate long memory by using the moving average representation of the cross-sectional aggregated process using the fast Fourier algorithm. See Vera-Valdes(2021) for details.

Arguments

x::Vector: time series

p::Float64: first parameter of the cross-sectional aggregated process

q::Float64: second parameter of the cross-sectional aggregated process, which is related to the fractional difference parameter d by q = 2(1-d)

Notes

q determines the long memory parameter of the cross-sectional aggregated process. The relation q = 2(1-d) holds, where d is the fractional difference parameter. We use autoregressive formulas to efficiently compute the coefficients of the moving average representation of the cross-sectional aggregated process.

Examples

julia> csadiff(randn(100,1),1.2,1.4)

LongMemory.csagen — Method

csagen(T::Int,p,q;μ=0,σ=1)

Generate a time series with long memory parameter q and length T using the cross-sectional aggregated process. FSee Vera-Valdes(2021) for details.

Arguments

T::Int: length of the time series

p::Float64: first parameter of the cross-sectional aggregated process

q::Float64: second parameter of the cross-sectional aggregated process, which is related to the fractional difference parameter d by q = 2(1-d)

Optional arguments

μ::Float64: mean of the time series

σ::Float64: standard deviation of the time series

Output

x::Vector: time series

Examples

julia> csagen(100,1.2,1.4)

LongMemory.csagen — Method

csagen(T::Int,N::Int,p,q;t=0.01;μ=0,σ=1)

Generate a time series with long memory parameter q and length T using the cross-sectional aggregation of 'N' AR(1) processes à la Granger (1980).

Arguments

T::Int: length of the time series

N::Int: number of AR(1) processes

p::Float64: first parameter of the cross-sectional aggregated process

q::Float64: second parameter of the cross-sectional aggregated process, which is related to the fractional difference parameter d by q = 2(1-d)

Optional arguments

t::Float64: taper length

μ::Float64: mean of the time series

σ::Float64: standard deviation of the time series

Notes

Multiple dispatch is used to generate the finite sample process if 'N' is included in the arguments.

Output

x::Vector: time series

Examples

julia> csagen(100,100,1.2,1.4)

LongMemory.csamle_est — Method

csa_mle_est(x::Array)

Computes the maximum likelihood estimate of the parameters p and q of the CSA process and the variance of the CSA process given the data x.

Arguments

x::Array: The data.

Notes

This function uses the Optim package to minimize the log-likelihood function.

Examples

julia> csa_mle_est(randn(100,1))

LongMemory.edmgen — Method

edmgen(T::Int,d; t=0.5, μ=0, σ=1)

Generate a time series with long memory parameter d and length T using the error duration model à la Parke (1999).

Arguments

T::Int: length of the time series

d::Float64: long memory parameter

Optional arguments

t::Float64: taper length

μ::Float64: mean of the time series

σ::Float64: standard deviation of the time series

Output

x::Vector: time series

Notes

The taper length t is the proportion of the time series that is pre-sampled to avoid the initial bias of the error duration model.

Examples

julia> edmgen(100,0.4)

LongMemory.exact_whittle_est — Method

exact_whittle_est(x::Array; m=0.5, l=0)

Estimate the long memory parameter of a time series x using the exact Whittle log-likelihood function. See Shimotsu and Phillips (2005) for details.

Arguments

x::Vector: time series

m∈(0,1)::Float64: taper final

l∈(0,1)::Float64: taper initial

Output

d::Float64: long memory parameter

Notes

The function considers the periodogram of the time series x for frequencies in the interval [T^l,T^m]. The zero frequency is always excluded. The condition m < l must hold. The default values of m and l are 0.5 and 0, respectively.

Examples

julia> exact_whittle_est(randn(100,1))

LongMemory.exact_whittle_llk — Method

exact_whittle_llk(d, x::Array; m=0.5, l=0)

Compute the exact Whittle log-likelihood function of a time series x for a given long memory parameter d. See Shimotsu and Phillips (2005) for details.

Arguments

d::Float64: long memory parameter

x::Vector: time series

m∈(0,1)::Float64: taper final

l∈(0,1)::Float64: taper initial

Output

Q::Float64: Whittle log-likelihood function

Notes

The function considers the periodogram of the time series x for frequencies in the interval [T^l,T^m]. The zero frequency is always excluded. The condition m < l must hold. The default values of m and l are 0.5 and 0, respectively.

Examples

julia> exact_whittle_llk(0.4,randn(100,1))

LongMemory.fi — Method

fi(T,d;μ=0,σ=1)

Generate a time series with long memory parameter d and length T using the fractional difference filter.

Arguments

T::Int: length of the time series

d::Float64: fractional difference parameter

Optional arguments

μ::Float64: mean of the time series

σ::Float64: standard deviation of the time series

Output

x::Vector: time series

Notes

Multiple dispatch is used for generation: If d is an integer, the function returns a time series with first or null difference. See fracdiff for details.

Examples

julia> fi(100,0.4)

LongMemory.fi_llk — Method

fi_llk(x::Array,d::Real)

Computes the log-likelihood of the fractional differenced process with parameter d given the data x.

Arguments

x::Array: The data.

d::Real: The fractional differencing parameter.

Notes

This function computes the concentrated log-likelihood function of the fractional differenced process with parameter d given the data x. The function is inspired by the arfima.Estimate() function in Ox; see Doornik (1999).

Examples

julia> fi_llk(randn(100,1), 0.4)

LongMemory.fi_survival_probs — Method

" fisurvivalprobs(N::Int,d)

Generate the survival probabilities of the error duration model à la Parke (1999).

Arguments

N::Int: length of the time series

d::Float64: fractional difference parameter

Output

p::Vector: survival probabilities

Notes

The survival probabilities are computed using the recursive formula p_{t+1} = p_t * (t + d - 1) / (t + 1 - d) to avoid numerical overflow.

Examples

julia> fi_survival_probs(100,0.4)

LongMemory.fi_var_matrix — Method

fi_var_matrix(T::Int, d::Real)

Constructs the autocovariance matrix of the fractional differenced process with parameter d at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

d::Real: The fractional differencing parameter.

Examples

julia> fi_var_matrix(10, 0.4)

LongMemory.fi_var_vals — Method

fi_var_vals(T::Int,d::Real)

Computes the autocovariance function of the fractional differenced process with parameter d at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

d::Real: The fractional differencing parameter.

Notes

This function uses the recursive formula for the autocovariance function of the fractional differenced process.

Examples

julia> fi_var_vals(10, 0.4)

LongMemory.fimle_est — Method

fimle_est(x::Array)

Computes the maximum likelihood estimate of the fractional differencing parameter and the variance of the fractional differenced process given the data x.

Arguments

x::Array: The data.

Notes

This function uses the Optim package to minimize the log-likelihood function. The function is inspired by the arfima.Estimate() function in Ox; see Doornik (1999).

Examples

julia> fimle_est(randn(100,1))

LongMemory.fracdiff — Method

fracdiff(x,d)

Compute the fractional difference of a time series x with fractional order d∈(-1/2,1/2).

Arguments

x::Vector: time series

d::Float64: fractional difference parameter

Notes

The function uses the fast Fourier transform to compute the convolution of the time series with the fractional difference filter. See Jensen and Nielsen (2014) for details. We use autoregressive formulas to efficiently compute the coefficients of the fractional difference filter.

Examples

julia> fracdiff(randn(100,1),0.4)

LongMemory.fracdiff — Method

fracdiff(x,d::Int)

Compute the first or null difference of a time series x. Multiple dispatch is used to return the same input or call the function diff from the Julia standard library if d=1 or d=0, respectively.

Arguments

x::Vector: time series

d::Int64: difference parameter

Examples

julia> fracdiff(randn(100,1),1)

LongMemory.gph_est — Method

gph_est(x::Array; m=0.5, l=0, br=0::Int)

Estimate the long memory parameter of a time series x using the log-periodogram estimator. See Geweke and Porter-Hudak (1983) and Andrews and Guggenberger (2003) for details.

Arguments

x::Vector: time series

m∈(0,1)::Float64: taper final

l∈(0,1)::Float64: taper initial

br::Int64: number of bias reduction terms

Notes

The function considers the periodogram of the time series x for frequencies in the interval [T^l,T^m]. The zero frequency is always excluded. The default values of m and l are 0.5 and 0, respectively. The condition m < l must hold.

The default value of br is 0 which returns the original GPH log-periodogram estimator.

Examples

julia> gph_est(randn(100,1))

LongMemory.gph_est_variance — Method

gph_est_variance(x::Array; m=0.5, l=0, br=0::Int)

Estimate the variance of the long memory parameter of a time series x using the log-periodogram estimator. See Geweke and Porter-Hudak (1983) and Andrews and Guggenberger (2003) for details.

Arguments

x::Vector: time series

Optional arguments

m∈(0,1)::Float64: taper final

br::Int64: number of bias reduction terms

Notes

Multiple dispatch is used for computation. If the first input is an integer, the function interprets it as the sample size; otherwise, it computes the sample size from the length of the time series.

Examples

julia> gph_est_variance(fi(100,0.4))

LongMemory.gph_est_variance — Method

gph_est_variance(T::Int; m=0.5, l=0, br=0::Int)

Estimate the variance of the long memory parameter of a time series of length T using the log-periodogram estimator. See Geweke and Porter-Hudak (1983) and Andrews and Guggenberger (2003) for details.

Arguments

T::Int: length of the time series

Optional arguments

m∈(0,1)::Float64: taper final

br::Int64: number of bias reduction terms

Notes

Multiple dispatch is used for computation. If the first input is an integer, the function interprets it as the sample size; otherwise, it computes the sample size from the length of the time series.

Examples

julia> gph_est_variance(100,0.4)

LongMemory.my_toeplitz — Method

my_toeplitz(coefs::Array)

Constructs a Toeplitz matrix from the given coefficients.

Arguments

coefs::Array: An array of coefficients.

Examples

julia> my_toeplitz([1, 2, 3])

LongMemory.periodogram — Method

periodogram(x::Array)

Compute the periodogram of a time series x using the fast Fourier transform.

Arguments

x::Vector: time series

Output

I_w::Vector: periodogram

w::Vector: Fourier frequencies

Examples

julia> periodogram(randn(100,1))

LongMemory.whittle_est — Method

whittle_est(x::Array; m=0.5, l=0)

Estimate the long memory parameter of a time series x using the Whittle log-likelihood function. See Künsch (1987) for details.

Arguments

x::Vector: time series

m∈(0,1)::Float64: taper final

l∈(0,1)::Float64: taper initial

Output

d::Float64: long memory parameter

Notes

The function considers the periodogram of the time series x for frequencies in the interval [T^l,T^m]. The zero frequency is always excluded. The condition m < l must hold. The default values of m and l are 0.5 and 0, respectively.

Examples

julia> whittle_est(randn(100,1))

LongMemory.whittle_est_variance — Method

whittle_est_variance(x::Array; m=0.5)

Estimate the variance of the estimator for the long memory parameter of a time series x using the Whittle log-likelihood function. See Künsch (1987) for details.

Arguments

x::Vector: time series

Optional arguments

m∈(0,1)::Float64: taper final

Notes

Multiple dispatch is used for computation. If the first input is an integer, the function interprets it as the sample size; otherwise, it computes the sample size from the length of the time series. The variance is the same as the one from using the exact Whittle log-likelihood function.

Examples

julia> whittle_est_variance(fi(100,0.4))

LongMemory.whittle_est_variance — Method

whittle_est_variance(T::Int;m=0.5

Estimate the variance of the estimator for the long memory parameter of a time series of length T using the Whittle log-likelihood function. See Künsch (1987) for details.

Arguments

T::Int: length of the time series

Optional arguments

m∈(0,1)::Float64: taper final

Notes

Multiple dispatch is used for computation. If the first input is an integer, the function interprets it as the sample size; otherwise, it computes the sample size from the length of the time series. The variance is the same as the one from using the exact Whittle log-likelihood function.

Examples

julia> whittle_est_variance(100,0.4)

LongMemory.whittle_llk — Method

whittle_llk(d, x::Array; m=0.5, l=0)

Compute the Whittle log-likelihood function of a time series x for a given long memory parameter d. See Künsch (1987) for details.

Arguments

d::Float64: long memory parameter

x::Vector: time series

m∈(0,1)::Float64: taper final

l∈(0,1)::Float64: taper initial

Output

Q::Float64: Whittle log-likelihood function

Notes

The function considers the periodogram of the time series x for frequencies in the interval [T^l,T^m]. The zero frequency is always excluded. The condition m < l must hold. The default values of m and l are 0.5 and 0, respectively.

Examples

julia> whittle_llk(0.4,randn(100,1))