Classic Estimator for the Hurst Effect

Functions to estimate the rescaled range (R/S) statistic and the Hurst coefficient. The estimator is based on the original work by Hurst (1951) and is included for historical reasons. The estimator is not recommended for practical purposes.

LongMemory.ClassicEstimators — Module

ClassicEstimators

This module contains functions to estimate them Hurst coefficient, closely related to the long memory parameter, of a time series using the rescaled range (R/S) statistic.

Author

J. Eduardo Vera-Valdés

LongMemory.ClassicEstimators.my_std — Method

my_std(x::Array)

Computes the sample standard deviation of a time series.

Arguments

x::Array: The time series.

Output

std::Real: The sample standard deviation.

Notes

This function divides by T instead of T-1.

Examples

julia> my_std(randn(100))

LongMemory.ClassicEstimators.rescaled_range — Method

rescaled_range(x::Array)

Computes the rescaled range (R/S) statistic of a time series.

Arguments

x::Array: The time series.

Output

RS::Array: The rescaled range statistic.

Examples

julia> rescaled_range(randn(100))

LongMemory.ClassicEstimators.rescaled_range_est — Method

rescaled_range_est(x::Array)

Estimates the Hurst coefficient of a time series using the rescaled range (R/S) statistic.

Arguments

x::Array: The time series.

Output

H::Real: The estimated Hurst coefficient.

Notes

This function uses the linear regression method on the log of the rescaled range to estimate the Hurst coefficient. The Hurst coefficient is related to the long memory parameter d by the formula H = d + 1/2.

Examples

julia> rescaled_range_est(randn(100))

Hurst, H.E. (1951). Long-term storage capacity of reservoirs. Transactions of the American Society of Civil Engineers, 116, 770-799.

Documentation for LongMemory.jl.