Structural Change Tests for Long Memory Processes

Functions to estimate structural changes in the long memory parameter of a time series.

LongMemory.StructuralChanges — Module

StructuralChanges

This module contains functions to estimate structural changes in models contaning long memory errors.

Author

J. Eduardo Vera-Valdés

LongMemory.StructuralChanges.lm_change_test — Method

lm_change_test(y::Array; uplim::Real = 0.15, lowlim::Real = 0.15)

Estimates the location of a structural change in a long memory model.

Arguments

y::Array: The series to be tested.

Optional arguments

uplim::Real = 0.15: The upper limit of the fraction of the series to be tested.

lowlim::Real = 0.15: The lower limit of the fraction of the series to be tested.

Output

τ::Int: The estimated location of the structural change.

Examples

julia> lm_change_test(randn(100))

Notes

The function estimates the location of a structural change in a long memory model. The function uses the Whittle estimator to estimate the long memory parameter. Then, it integrates the series and computes the t-statistics of the OLS regression of the integrated series on the starred series. The function computes the forward and backward t-statistics and returns the location of the maximum squared t-statistic. Hence, it is robust to the direction of the change.

References

Martins and Rodrigues (2014), "Testing for persistence change in fractionally integrated models: An application to world inflation rates", Computational Statistics and Data Analysis, 76.

LongMemory.StructuralChanges.simple_ols — Method

simple_ols(y::Array, x::Array)

Computes the t-statistics of the OLS regression of y on x.

Arguments

y::Array: The dependent variable.

x::Array: The independent variable.

Output

tstat::Array: The t-statistic of the regression.

Notes

The function should only be used internally. It is not exported. It is used to compute the t-statistics of the OLS regression in the lm_change_test function. Hence, it does not adds intercepts to the regression nor it returns the coefficient.

LongMemory.StructuralChanges.starred_var — Method

starred_var(x::Array)

Computes the padded series x with detrending coefficients.

Arguments

x::Array: The series to be detrended.

Output

dx::Array: The detrended series.

Documentation for LongMemory.jl.