Parametric Estimators for Long Memory

Functions to estimate time series long memory models by parametric methods. Of particular interest is the ARFIMA model.

LongMemory.ParametricEstimators — Module

ParametricEstimators

This module contains functions for estimating the parameters of the fractional differenced process à la ARFIMA and the CSA process.

Author

J. Eduardo Vera-Valdés

LongMemory.ParametricEstimators.csa_llk — Method

csa_llk(p::Real, q::Real, x::Array)

Computes the log-likelihood of the CSA process with parameters p and q given the data x.

Arguments

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

x::Array: The data.

Notes

This function computes the concentrated log-likelihood function of the CSA process with parameters p and q given the data x.

Examples

julia> csa_llk(1.4, 1.8, randn(100,1))

LongMemory.ParametricEstimators.csa_var_matrix — Method

csa_var_matrix(T::Int, d::Real)

Constructs the autocovariance matrix of the CSA process with parametersp and q at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

Examples

julia> fi_var_matrix(10, 1.4, 1.8)

LongMemory.ParametricEstimators.csa_var_vals — Method

csa_var_vals(T::Int, p::Real, q::Real)

Computes the autocovariance function of the CSA process with parameters p and q at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

p::Real: The first parameter of the CSA process.

q::Real: The second parameter of the CSA process.

Notes

This function uses the recursive formula for the autocovariance function of the CSA process.

Examples

julia> csa_var_vals(20, 0.4, 0.6)

LongMemory.ParametricEstimators.csamle_est — Method

csa_mle_est(x::Array)

Computes the maximum likelihood estimate of the parameters p and q of the CSA process and the variance of the CSA process given the data x.

Arguments

x::Array: The data.

Notes

This function uses the Optim package to minimize the log-likelihood function.

Examples

julia> csa_mle_est(randn(100,1))

LongMemory.ParametricEstimators.fi_llk — Method

fi_llk(x::Array,d::Real)

Computes the log-likelihood of the fractional differenced process with parameter d given the data x.

Arguments

x::Array: The data.

d::Real: The fractional differencing parameter.

Notes

This function computes the concentrated log-likelihood function of the fractional differenced process with parameter d given the data x. The function is inspired by the arfima.Estimate() function in Ox; see Doornik (1999).

Examples

julia> fi_llk(randn(100,1), 0.4)

LongMemory.ParametricEstimators.fi_var_matrix — Method

fi_var_matrix(T::Int, d::Real)

Constructs the autocovariance matrix of the fractional differenced process with parameter d at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

d::Real: The fractional differencing parameter.

Examples

julia> fi_var_matrix(10, 0.4)

LongMemory.ParametricEstimators.fi_var_vals — Method

fi_var_vals(T::Int,d::Real)

Computes the autocovariance function of the fractional differenced process with parameter d at lags 0, 1, ..., T-1.

Arguments

T::Int: The number of lags to compute.

d::Real: The fractional differencing parameter.

Notes

This function uses the recursive formula for the autocovariance function of the fractional differenced process.

Examples

julia> fi_var_vals(10, 0.4)

LongMemory.ParametricEstimators.fimle_est — Method

fimle_est(x::Array)

Computes the maximum likelihood estimate of the fractional differencing parameter and the variance of the fractional differenced process given the data x.

Arguments

x::Array: The data.

Notes

This function uses the Optim package to minimize the log-likelihood function. The function is inspired by the arfima.Estimate() function in Ox; see Doornik (1999).

Examples

julia> fimle_est(randn(100,1))

LongMemory.ParametricEstimators.my_toeplitz — Method

my_toeplitz(coefs::Array)

Constructs a Toeplitz matrix from the given coefficients.

Arguments

coefs::Array: An array of coefficients.

Examples

julia> my_toeplitz([1, 2, 3])

Documentation for LongMemory.jl.